requirements.txt
google-cloud-bigquery>=3.10.0,<3.11.0
google-cloud-storage>=2.10.0,<2.11.0
pandas>=2.1.0,<2.2.0
pandas-gbq>=0.19.0,<0.20.0
